NPC indication & sides look
This commit is for the NPC behavior, this consists of a script for the NPC to turn to the side that the player character approach's from. i used the prefab characters and used then for construction of the NPC turning mechanic. I also added a script that would be used on all NPC characters as it is for the indication of the character itself. when the player approches the NPC an icon is to pop up above the NPC.