Dialogue & Ink intergration
In this commit, I make a number of scripts that deal with the dialogue management as well as a temporary script for the player inputs. The scripts are NCPmovement, DialogueScript, InputsHandler. I renamed some scripts so that we can differentiate the temporary scripts(those that will be deleted later on). I also fixed some bugs I was having and the code is functionally as of this commit. I added a button as well that will continue to read the rest of the dialogue line by line, from the INK files that will hold the dialogue.