Bug fixing and file sorting
In this commit I fixed some of the visual bugs that I had when displaying the dialogue as well as the NPC indicator. I still have a bug for the dialogue choices that i couldn't fix, but i should have it fixed by my next commit.  I added a temporary banner for the NPC that would appear when the player character is near.