Bug update
I sorted all the issues I had with the bugs. The flipping message bug has been fixed, this occured when a player character was next to the NPC, when the NPC looked to the player the text prompt for the NPC would also flip backwards. I fixed that. I also fixed the Null reference exception that was popping up after every dialogue ended. I just had to replace the null value with a ' "" ', in the code and it made sure to set the dialogue to the empty quotes and thus clearing it while not setting it to null completely.