Dialogue files update
In this commit I made all the inkle files that will be used on all the NPCs. essentially what is left to do is add them to the NPCs that will be used and test is everything is functional. I have personally tested the Inkle files and they are all working, no bugs have been detected. most of the things that are left are just the cosmetic things like fonts and the color of the displayed text.