Dialogue image display and name display
In this commit I have made sure to make the dialogue display an image and the name of the person that is speaking at any given moment. I changed the inkle dialogue files for the beggar to make sure that I use tags in order to make the swapping of names and images in the animation tile smoother and on time.