-simar dev: Made changes to the signup screen and added code to save initial player data.
I) Made changes to the signup up panel to the save the email in the database once the verification link has been sent, so that the user doesn't get the error message of "email already in use" for an unregistered email.
II) Made changes in the code so that once the link is verified the user can click signup again to enter the game screen instead of going to the login screen to start their game.
iii) Once the user is verified and signs up their details are added to the realtime database under their unique user ID.
iv) Initial game data for a player is added to the realtime database once they start the game .