Updates for the Save, Restart ,Quit and the continue button
Made changes to the code to save the player money and favourability in the realtime database. When the player clicks save or quit in the pause menu their progress gets saved in the database. When the player presses continue they start the game back with their saved data. When they click restart they can restart the game.